refactor(reco-io): extract VAAPI hw upload into hw_upload.rs
Move HardwareUpload + staging_pixel_format out of encoder.rs into a
dedicated ffmpeg/hw_upload.rs, isolating all the VAAPI hwframe unsafe
FFI in one file. Encapsulate the hw_frames_ctx attach as
attach_to_encoder so encoder.rs no longer touches the raw frames_ref.
No behavior change.
